updatefcmtoken moved to viewmodel
